+++ packages/atomic/src/components/common/smart-snippets/atomic-smart-snippet-expandable-answer/atomic-smart-snippet-expandable-answer.tsx
@@ -9,10 +9,10 @@ import {
   Event,
   EventEmitter,
 } from '@stencil/core';
-import ArrowDown from '../../../images/arrow-down.svg';
-import {listenOnce} from '../../../utils/event-utils';
-import {InitializeBindings} from '../../../utils/initialization-utils';
-import {Bindings} from '../atomic-search-interface/atomic-search-interface';
+import ArrowDown from '../../../../images/arrow-down.svg';
+import {listenOnce} from '../../../../utils/event-utils';
+import {InitializeBindings} from '../../../../utils/initialization-utils';
+import {AnyBindings} from '../../interface/bindings';
 
 /**
  * @internal
@@ -23,7 +23,7 @@ import {Bindings} from '../atomic-search-interface/atomic-search-interface';
   shadow: true,
 })
 export class AtomicSmartSnippetExpandableAnswer {
-  @InitializeBindings() public bindings!: Bindings;
+  @InitializeBindings() public bindings!: AnyBindings;
   public error!: Error;
   @Element() public host!: HTMLElement;