+++ packages/atomic/src/components/search/facets/atomic-rating-facet/atomic-rating-facet.tsx
@@ -62,6 +62,7 @@ import {initializePopover} from '../atomic-popover/popover-type';
  * @part value-checkbox - The facet value checkbox, available when display is 'checkbox'.
  * @part value-checkbox-checked - The checked facet value checkbox, available when display is 'checkbox'.
  * @part value-checkbox-label - The facet value checkbox clickable label, available when display is 'checkbox'.
+ * @part value-checkbox-icon - The facet value checkbox icon, available when display is 'checkbox'.
  * @part value-link - The facet value when display is 'link'.
  * @part value-link-selected - The selected facet value when display is 'link'.
  */