+++ packages/headless/src/controllers/core/sort/headless-core-sort.test.ts
@@ -1,4 +1,4 @@
-import {configuration, sortCriteria} from '../../../app/reducers';
+import {configuration} from '../../../app/common-reducers';
 import {updatePage} from '../../../features/pagination/pagination-actions';
 import {
   buildRelevanceSortCriterion,
@@ -12,6 +12,7 @@ import {
   registerSortCriterion,
   updateSortCriterion,
 } from '../../../features/sort-criteria/sort-criteria-actions';
+import {sortCriteriaReducer as sortCriteria} from '../../../features/sort-criteria/sort-criteria-slice';
 import {
   MockSearchEngine,
   buildMockSearchAppEngine,