+++ packages/headless/src/controllers/facets/category-facet/headless-category-facet.test.ts
@@ -1,9 +1,4 @@
-import {
-  categoryFacetSearchSet,
-  categoryFacetSet,
-  configuration,
-  search,
-} from '../../../app/reducers';
+import {configuration} from '../../../app/common-reducers';
 import {updateFacetOptions} from '../../../features/facet-options/facet-options-actions';
 import {
   registerCategoryFacet,
@@ -12,15 +7,18 @@ import {
   updateCategoryFacetNumberOfValues,
   updateCategoryFacetSortCriterion,
 } from '../../../features/facets/category-facet-set/category-facet-set-actions';
+import {categoryFacetSetReducer as categoryFacetSet} from '../../../features/facets/category-facet-set/category-facet-set-slice';
 import {defaultCategoryFacetOptions} from '../../../features/facets/category-facet-set/category-facet-set-slice';
 import {
   CategoryFacetRequest,
   CategoryFacetSortCriterion,
 } from '../../../features/facets/category-facet-set/interfaces/request';
+import {categoryFacetSearchSetReducer as categoryFacetSearchSet} from '../../../features/facets/facet-search-set/category/category-facet-search-set-slice';
 import {
   executeSearch,
   fetchFacetValues,
 } from '../../../features/search/search-actions';
+import {searchReducer as search} from '../../../features/search/search-slice';
 import {SearchAppState} from '../../../state/search-app-state';
 import {
   buildMockSearchAppEngine,