+++ packages/quantic/cypress/page-objects/search.ts
@@ -275,15 +275,18 @@ export function mockSearchWithoutAnyFacetValues(useCase: string) {
   }).as(InterceptAliases.Search.substring(1));
 }
 
-export function mockSearchWithSmartSnippet(smartSnippetOptions: {
-  question: string;
-  answer: string;
-  title: string;
-  uri: string;
-  permanentId: string;
-}) {
+export function mockSearchWithSmartSnippet(
+  smartSnippetOptions: {
+    question: string;
+    answer: string;
+    title: string;
+    uri: string;
+    permanentId: string;
+  },
+  useCase?: string
+) {
   const {question, answer, title, uri, permanentId} = smartSnippetOptions;
-  cy.intercept(routeMatchers.search, (req) => {
+  cy.intercept(getRoute(useCase), (req) => {
     req.continue((res) => {
       res.body.questionAnswer = {
         answerFound: true,
@@ -293,6 +296,7 @@ export function mockSearchWithSmartSnippet(smartSnippetOptions: {
           contentIdKey: 'permanentid',
           contentIdValue: permanentId,
         },
+        relatedQuestions: [],
       };
       res.body.results = [
         {
@@ -309,11 +313,12 @@ export function mockSearchWithSmartSnippet(smartSnippetOptions: {
   }).as(InterceptAliases.Search.substring(1));
 }
 
-export function mockSearchWithoutSmartSnippet() {
-  cy.intercept(routeMatchers.search, (req) => {
+export function mockSearchWithoutSmartSnippet(useCase?: string) {
+  cy.intercept(getRoute(useCase), (req) => {
     req.continue((res) => {
       res.body.questionAnswer = {
         answerFound: false,
+        relatedQuestions: [],
       };
       res.send();
     });
@@ -330,12 +335,17 @@ export function mockSearchWithSmartSnippetSuggestions(
       contentIdKey: string;
       contentIdValue: string;
     };
-  }>
+  }>,
+  useCase?: string
 ) {
-  cy.intercept(routeMatchers.search, (req) => {
+  cy.intercept(getRoute(useCase), (req) => {
     req.continue((res) => {
       res.body.questionAnswer = {
         relatedQuestions: relatedQuestions,
+        documentId: {
+          contentIdKey: 'permanentid',
+          contentIdValue: 'example permanentId',
+        },
       };
       res.body.results = relatedQuestions.map(({title, uri, documentId}) => ({
         uri,
@@ -350,8 +360,8 @@ export function mockSearchWithSmartSnippetSuggestions(
   }).as(InterceptAliases.Search.substring(1));
 }
 
-export function mockSearchWithoutSmartSnippetSuggestions() {
-  cy.intercept(routeMatchers.search, (req) => {
+export function mockSearchWithoutSmartSnippetSuggestions(useCase?: string) {
+  cy.intercept(getRoute(useCase), (req) => {
     req.continue((res) => {
       res.body.questionAnswer = {
         relatedQuestions: [],